fix flatten for oidc config
The current implementation for flatten of the kube API server is not including the OIDC config. This should be fixed with this change. In addition the schema for the `signing_algs` and `required_claims` was not correct. based on the Gardener API this needs to be a Set/ Map instead of a string.